Consolidate tx quering logic into a single file
